fix datatype for 64 bit architectures (cosmetic for now)
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@41482 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/httpd/httpd.c
@@ -782,7 +782,7 @@ static void *handle_request(void *arg)
 	char *file = NULL;
 	int len;
 	struct mime_handler *handler = NULL;;
-	unsigned long content_length = 0, flags;
+	size_t content_length = 0, flags;
 	char *line;
 	long method_type;
 	conn_fp->p = &global_vars;
